Support various access strides
Now a user can set the stride size of the object.
This is useful when comparing sequential access and random access,
because the nr_page_faults is very high especially in random access by
the default stride (single-word; 4 Byte). To see the effectiveness of
swapping (read-ahead or prediction), it is better to use page-size
stride.

Signed-off-by: Yunjae Lee <lyj7694@gmail.com>